feat: Changed files according to the suggestion
